Cleaning up main.py
Changing most of the variable names in the checkColUnique function. Also added a lot of comments. Altogether the code is much, much easier to understand.

Also streamlined a few things. A couple of the checks in checkColUnique were running many more times than was necessary. Overall runtime should be drastically reduced now.

Still need to update checkRowUnique, checkBlockUnique, and the guess function. The first two should be pretty straightforward, as they're only variations on checkColUnique. The guess function, on the other hand, needs a serious overhaul. So it may be a few days for that one.