Updating main
Cleaned up variable names and comments for checkRowUnique and checkBlockUnique. Streamlined code for checkRowUnique and some of the main loop. Notably, some functions were being called in loops (i.e. for x in range (0, 81) call printSudoku(x)). A lot of these loops have been implemented into the functions, and the main loop is now much cleaner.

The guess function is next to be worked on. Gonna hold off on merging this to the Master branch for now due to the guess function rewrite being incomplete, making that part of the code dysfunctional.